+++ src/main/java/com/hojunara/web/dto/response/DetailedPropertyPostDto.java
@@ -49,7 +49,7 @@ public class DetailedPropertyPostDto {
 
     private Timestamp createdAt;
 
-    private double viewCounts;
+    private Long viewCounts;
 
     private Long likeCounts;
