+++ src/main/java/com/hojunara/web/dto/response/DetailedStudyPostDto.java
@@ -30,7 +30,7 @@ public class DetailedStudyPostDto {
 
     private Timestamp createdAt;
 
-    private double viewCounts;
+    private Long viewCounts;
 
     private List<Map<String, String>> blogContents;
